+++ CHANGELOG.md
@@ -40,6 +40,7 @@ This version also changes the public API scope of the project:
 * Support GZIP compressed datasets
 * `eval` run type accepts `--features_file` and `--labels_file` to evaluate files other than the ones defined in the YAML configuration
 * Accept `with_alignments: soft` to output soft alignments in inference or scoring
+* `dropout` can be configured in the YAML configuration to override the model values
 
 ### Fixes and improvements